Fix bug causing runner to panic in case a task fails

+++ examples/workflows/mapreduce/mapreduce.go
@@ -10,8 +10,8 @@ import (
 
 // RootTask is the entry point that spawns all map and reduce tasks.
 type RootTask struct {
-	NumMapTasks    int
-	NumReduceTasks int
+	NumMapTasks    uint64
+	NumReduceTasks uint64
 }
 
 func (t *RootTask) Execute(ctx context.Context) error {
@@ -41,32 +41,32 @@ func (t *RootTask) Execute(ctx context.Context) error {
 		}
 	}
 
-	slog.Info("Submitted reduce tasks", slog.Int("count", t.NumReduceTasks))
+	slog.Info("Submitted reduce tasks", slog.Uint64("count", t.NumReduceTasks))
 
 	reduceProgress := workflows.Progress("reduce")
-	_ = reduceProgress.Add(ctx, uint64(t.NumReduceTasks))
+	_ = reduceProgress.Add(ctx, t.NumReduceTasks)
 
 	return nil
 }
 
 // MapTask processes a single item of input data.
 type MapTask struct {
-	Index int
+	Index uint64
 }
 
 func (t *MapTask) Execute(ctx context.Context) error {
-	slog.Info("Executing map task", slog.Int("index", t.Index))
+	slog.Info("Executing map task", slog.Uint64("index", t.Index))
 	_ = workflows.Progress("map").Done(ctx, 1)
 	return nil
 }
 
 // ReduceTask aggregates results from map tasks.
 type ReduceTask struct {
-	ReducerIndex int
+	ReducerIndex uint64
 }
 
 func (t *ReduceTask) Execute(ctx context.Context) error {
-	slog.Info("Executing reduce task", slog.Int("reducer_index", t.ReducerIndex))
+	slog.Info("Executing reduce task", slog.Uint64("reducer_index", t.ReducerIndex))
 	_ = workflows.Progress("reduce").Done(ctx, 1)
 	return nil
 }
